# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Message Hub API
  description: |-
    API for the Message Hub service, providing capabilities to publish and consume messages from Kafka and RabbitMQ.
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://message-hub.example.com
    description: Production server
...
paths:
  /api/messages:
    post:
      tags:
        - messages
      summary: Publish a message
      description: Publishes a message to Kafka or RabbitMQ
      operationId: publishMessage
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishMessageRequest'
      responses:
        '200':
          description: Message published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishMessageResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/history:
    get:
      tags:
        - messages
      summary: Get message history
      description: Retrieves message history for a destination
      operationId: getMessageHistory
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: destination
          in: query
          description: Message destination (topic or queue)
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of messages to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Number of messages to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Message history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageHistoryResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/destinations:
    get:
      tags:
        - messages
      summary: Get available destinations
      description: Retrieves available topics and queues
      operationId: getDestinations
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Destinations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/schemas:
    get:
      tags:
        - messages
      summary: Get message schemas
      description: Retrieves available message schemas
      operationId: getMessageSchemas
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Schemas retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/JsonSchema'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/{destination}:
    delete:
      tags:
        - messages
      summary: Delete messages
      description: Deletes messages from a destination (admin only)
      operationId: deleteMessages
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: destination
          in: path
          description: Message destination (topic or queue)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    PublishMessageRequest:
      type: object
      required:
        - destination
        - message
      properties:
        destination:
          type: string
          description: Destination topic or queue (use rabbitmq:// or kafka:// prefix to specify protocol)
          example: "kafka://orders"
        message:
          description: Message content
          example: {"orderId": "12345", "customer": "John Doe"}
        messageType:
          type: string
          description: Message type for schema validation
          example: "order"
        routingKey:
          type: string
          description: Routing key for RabbitMQ
          example: "new.order"
        headers:
          type: object
          description: Message headers
          example: {"priority": "high"}
        options:
          type: object
          description: Additional options
          example: {"broadcast": true, "directQueue": false}
    PublishMessageResponse:
      type: object
      properties:
        success:
          type: boolean
        transactionId:
          type: string
        destination:
          type: string
        protocol:
          type: string
          enum: [kafka, rabbitmq]
        timestamp:
          type: string
          format: date-time
    MessageHistoryResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              destination:
                type: string
              timestamp:
                type: string
                format: date-time
              message:
                type: object
              headers:
                type: object
        pagination:
          type: object
          properties:
            totalCount:
              type: integer
            limit:
              type: integer
            offset:
              type: integer
            hasMore:
              type: boolean
    DestinationsResponse:
      type: object
      properties:
        kafka:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              messageCount:
                type: integer
              partitions:
                type: integer
        rabbitmq:
          type: object
          properties:
            queues:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  messageCount:
                    type: integer
                  consumers:
                    type: integer
            exchanges:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  type:
                    type: string
                  queues:
                    type: array
                    items:
                      type: string
    JsonSchema:
      type: object
      properties:
        type:
          type: string
        required:
          type: array
          items:
            type: string
        properties:
          type: object
          additionalProperties:
            type: object
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        statusCode:
          type: integer
        message:
          type: string
        details:
          type: object
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT